pyproject.toml
[project]
name = "pantrychef"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",        # async Postgres driver
    "pydantic-settings>=2.2.0",
    "python-jose[cryptography]>=3.3.0",  # JWT
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",           # for TestClient in tests
    "openai>=1.30.0",
    "celery[redis]>=5.4.0",
    "boto3>=1.34.0",           # AWS SES
    "pydantic[email]>=2.0.0",
    "greenlet>=3.0.0",
    "bcrypt==4.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",             # linter + formatter
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]      # errors, pyflakes, import sorting